style: apply lint

+++ contracts/library/LibQuestryPlatform.sol
@@ -14,7 +14,6 @@ library LibQuestryPlatform {
   bytes4 public constant ERC20_PAYMENT_MODE = bytes4(keccak256("ERC20"));
 
   //Role difinition
-  
 
   struct AllocateArgs {
     IPJManager pjManager;
@@ -50,24 +49,26 @@ library LibQuestryPlatform {
     bytes4 algorithm; // calculation algorithm for the board.
     bytes args; // arguments for the calculation algorithm.
   }
-  
+
   // ---- EIP712 ----
   bytes32 private constant AllOCATE_TYPEHASH =
     keccak256(
       "AllocateArgs(address pjManager,bytes4 paymentMode,address paymentToken,address board,CalculateDispatchArgs calculateArgs,address[] updateNeededPools,address[] contributePoolOwner,uint256 pjnonce)CalculateDispatchArgs(bytes4 algorithm,bytes args)"
     );
-  
+
   bytes32 private constant CALCURATEDISPATCHARGS_TYPEHASH =
-    keccak256(
-      "CalculateDispatchArgs(bytes4 algorithm,bytes args)"
-    );
+    keccak256("CalculateDispatchArgs(bytes4 algorithm,bytes args)");
 
   /**
    * @dev Prepares keccak256 hash for Allocate
    *
    * @param _allocateargs LibQuestryPlatform.AllocateArgs
    */
-  function _hashAllocate(AllocateArgs calldata _allocateargs) internal pure returns (bytes32) {
+  function _hashAllocate(AllocateArgs calldata _allocateargs)
+    internal
+    pure
+    returns (bytes32)
+  {
     return
       keccak256(
         abi.encode(
@@ -89,7 +90,9 @@ library LibQuestryPlatform {
    *
    * @param _calculatedispatchargs LibQuestryPlatform.CalculateDispatchArgs
    */
-  function _hashCalculateDispatchArgs(CalculateDispatchArgs calldata _calculatedispatchargs) internal pure returns (bytes32) {
+  function _hashCalculateDispatchArgs(
+    CalculateDispatchArgs calldata _calculatedispatchargs
+  ) internal pure returns (bytes32) {
     return
       keccak256(
         abi.encode(
@@ -103,10 +106,9 @@ library LibQuestryPlatform {
   /**
    * @dev validation check
    */
-  function _checkParameterForAllocation(
-    AllocateArgs calldata _allocateargs
-  ) internal {
-    
+  function _checkParameterForAllocation(AllocateArgs calldata _allocateargs)
+    internal
+  {
     //pjmaneger validation
     require(
       address(_allocateargs.pjManager) != address(0),
@@ -117,9 +119,9 @@ library LibQuestryPlatform {
       bytes4(_allocateargs.paymentMode) != bytes4(0),
       "LibQuestryPlatform: PaymentMode is Invalid"
     );
-    
+
     //Whitelist check
-    if(_allocateargs.paymentMode == ERC20_PAYMENT_MODE){
+    if (_allocateargs.paymentMode == ERC20_PAYMENT_MODE) {
       //paymnet Token Address validation
       require(
         address(_allocateargs.paymentToken) != address(0),
@@ -128,9 +130,7 @@ library LibQuestryPlatform {
       // Whitelist check
       // EM: BuyOrder paymentToken not whitelisted
       require(
-          _allocateargs.pjManager.isWhitelisted(
-          _allocateargs.paymentToken
-        ),
+        _allocateargs.pjManager.isWhitelisted(_allocateargs.paymentToken),
         "LibQuestryPlatform: Is not PJ Whitelist token"
       );
     }
@@ -143,18 +143,19 @@ library LibQuestryPlatform {
     /**
      * @dev : Calcurator args is validation skip
      */
-  
+
     //calcuration pool element check &ContributePool Owner check
     require(
       _allocateargs.updateNeededPools.length > 0,
-      'LibQuestryPlatform: contribution pool is zero'
+      "LibQuestryPlatform: contribution pool is zero"
     );
     require(
-      _allocateargs.updateNeededPools.length == _allocateargs.contributePoolOwner.length,
+      _allocateargs.updateNeededPools.length ==
+        _allocateargs.contributePoolOwner.length,
       "LibQuestryPlatform: array element is differnt"
     );
     //Contributioonpool check
-    for(uint idx = 0;idx < _allocateargs.updateNeededPools.length;idx++){
+    for (uint256 idx = 0; idx < _allocateargs.updateNeededPools.length; idx++) {
       require(
         address(_allocateargs.updateNeededPools[idx]) != address(0),
         "LibQuestryPlatform: contribution pool address is invalid"
@@ -164,8 +165,9 @@ library LibQuestryPlatform {
         "LibQuestryPlatform: contribution pool owner address is invalid"
       );
     }
-    require(_allocateargs.pjManager.getNonce() == _allocateargs.pjnonce,"LibQuestryPlatform: message nonce is different from on-chain nonce");
-    
+    require(
+      _allocateargs.pjManager.getNonce() == _allocateargs.pjnonce,
+      "LibQuestryPlatform: message nonce is different from on-chain nonce"
+    );
   }
-
 }
